Move Hash specs to rely less on shared examples
This is mostly a mechanical translation. A few things of interest:

Two shared specs weren't actually shared. `Hash#index` used to be an alias for `Hash#key` but was removed at some point. `Hash#values_at` had shared specs which never were actually shared.

Tests for `Hash#[]=` used `send`. I explicitly added back a test usign `send` to assert the return value.

I cleaned up requires, each test file can be run in isolation.

For `../mspec/bin/mspec core/hash`:

```
Before:
69 files, 633 examples, 1176 expectations, 0 failures, 0 errors, 0 tagged

After:
69 files, 558 examples, 1026 expectations, 0 failures, 0 errors, 0 tagged
```

So 75 fewer specs, 150 fewer assertions
