+++ core/hash/each_spec.rb
@@ -1,11 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/iteration'
-require_relative 'shared/each'
-require_relative '../enumerable/shared/enumeratorized'
 
 describe "Hash#each" do
-  it_behaves_like :hash_each, :each
-  it_behaves_like :hash_iteration_no_block, :each
-  it_behaves_like :enumeratorized_with_origin_size, :each, { 1 => 2, 3 => 4, 5 => 6 }
+  it "is an alias of Hash#each_pair" do
+    Hash.instance_method(:each).should == Hash.instance_method(:each_pair)
+  end
 end
