+++ core/hash/has_key_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/key'
 
 describe "Hash#has_key?" do
-  it_behaves_like :hash_key_p, :has_key?
+  it "is an alias of Hash#include?" do
+    Hash.instance_method(:has_key?).should == Hash.instance_method(:include?)
+  end
 end
