+++ core/hash/length_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/length'
 
 describe "Hash#length" do
-  it_behaves_like :hash_length, :length
+  it "is an alias of Hash#size" do
+    Hash.instance_method(:size).should == Hash.instance_method(:length)
+  end
 end
