+++ core/hash/member_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/key'
 
 describe "Hash#member?" do
-  it_behaves_like :hash_key_p, :member?
+  it "is an alias of Hash#include?" do
+    Hash.instance_method(:member?).should == Hash.instance_method(:include?)
+  end
 end
