--- core/hash/shared/values_at.rb
@@ -1,9 +1,0 @@
-describe :hash_values_at, shared: true do
-  it "returns an array of values for the given keys" do
-    h = { a: 9, b: 'a', c: -10, d: nil }
-    h.send(@method).should.is_a?(Array)
-    h.send(@method).should == []
-    h.send(@method, :a, :d, :b).should.is_a?(Array)
-    h.send(@method, :a, :d, :b).should == [9, nil, 'a']
-  end
-end
