+++ core/hash/size_spec.rb
@@ -1,7 +1,14 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/length'
 
 describe "Hash#size" do
-  it_behaves_like :hash_length, :size
+  it "returns the number of entries" do
+    { a: 1, b: 'c' }.size.should == 2
+    h = { a: 1, b: 2 }
+    h[:a] = 2
+    h.size.should == 2
+    { a: 1, b: 1, c: 1 }.size.should == 3
+    {}.size.should == 0
+    Hash.new(5).size.should == 0
+    Hash.new { 5 }.size.should == 0
+  end
 end
