+++ core/hash/store_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/store'
 
 describe "Hash#store" do
-  it_behaves_like :hash_store, :store
+  it "is an alias of Hash#[]=" do
+    Hash.instance_method(:store).should == Hash.instance_method(:[]=)
+  end
 end
