+++ core/hash/values_at_spec.rb
@@ -1,7 +1,11 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/values_at'
 
 describe "Hash#values_at" do
-  it_behaves_like :hash_values_at, :values_at
+  it "returns an array of values for the given keys" do
+    h = { a: 9, b: 'a', c: -10, d: nil }
+    h.values_at.should.is_a?(Array)
+    h.values_at.should == []
+    h.values_at(:a, :d, :b).should.is_a?(Array)
+    h.values_at(:a, :d, :b).should == [9, nil, 'a']
+  end
 end
